+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.8.4
 
+[ADDED]
+
+- Fuzzy searching for dropdowns
+
 [FIXED]
 
 - Downloading files while filtering by a dropdown question now works